+++ symbolic-debuginfo/src/function_builder.rs
@@ -29,6 +29,7 @@ pub struct FunctionBuilder<'s> {
     /// The lines, in any order. They will be sorted in `finish()`. These record specify locations
     /// at the innermost level of the inline stack at the line record's address.
     lines: Vec<LineInfo<'s>>,
+    max_inline_depth: Option<u32>,
 }
 
 impl<'s> FunctionBuilder<'s> {
@@ -41,7 +42,16 @@ impl<'s> FunctionBuilder<'s> {
             size,
             inlinees: BinaryHeap::new(),
             lines: Vec::new(),
+            max_inline_depth: None,
         }
+    }
+
+    /// Sets the maximum inline nesting depth to process.
+    ///
+    /// Inline records nested deeper than this are dropped.
+    pub fn max_inline_depth(mut self, max_inline_depth: Option<u32>) -> Self {
+        self.max_inline_depth = max_inline_depth;
+        self
     }
 
     /// Add an inlinee record. This method can be called in any order.
@@ -56,8 +66,9 @@ impl<'s> FunctionBuilder<'s> {
         call_file: FileInfo<'s>,
         call_line: u64,
     ) {
-        // An inlinee that starts before the function is obviously bogus.
-        if address < self.address {
+        // An inlinee that starts before the function is obviously bogus same for an inlinee that
+        // has a depth deeper than the limit.
+        if address < self.address || depth > self.max_inline_depth.unwrap_or(u32::MAX) {
             return;
         }
 
@@ -109,6 +120,7 @@ impl<'s> FunctionBuilder<'s> {
             size,
             inlinees,
             mut lines,
+            max_inline_depth: _,
         } = self;
 
         let inlinees = ensure_proper_nesting(inlinees);